+++ Buildenator/MakeBuilderAttributeInternal.cs
@@ -3,8 +3,16 @@ using Microsoft.CodeAnalysis;
 
 namespace Buildenator;
 
-internal readonly struct MakeBuilderAttributeInternal
+internal readonly struct MakeBuilderAttributeInternal(
+    INamedTypeSymbol typeForBuilder,
+    string? buildingMethodsPrefix,
+    bool? staticCreator,
+    NullableStrategy? nullableStrategy,
+    bool? generateMethodsForUnreachableProperties,
+    bool? implicitCast,
+    string? staticFactoryMethodName)
 {
+
     public MakeBuilderAttributeInternal(AttributeData attribute)
         : this(
             (INamedTypeSymbol)attribute.ConstructorArguments[0].Value!,
@@ -14,25 +22,17 @@ internal readonly struct MakeBuilderAttributeInternal
                 ? null
                 : (NullableStrategy)attribute.ConstructorArguments[3].Value!,
             (bool?)attribute.ConstructorArguments[4].Value,
-            (bool?)attribute.ConstructorArguments[5].Value)
+            (bool?)attribute.ConstructorArguments[5].Value,
+            (string?)attribute.ConstructorArguments[6].Value)
     {
 
-		}
+    }
 
-    public MakeBuilderAttributeInternal(INamedTypeSymbol typeForBuilder, string? buildingMethodsPrefix, bool? staticCreator, NullableStrategy? nullableStrategy, bool? generateMethodsForUnreachableProperties, bool? implicitCast)
-    {
-            TypeForBuilder = typeForBuilder;
-            BuildingMethodsPrefix = buildingMethodsPrefix;
-            DefaultStaticCreator = staticCreator;
-            ImplicitCast = implicitCast;
-            NullableStrategy = nullableStrategy;
-            GenerateMethodsForUnreachableProperties = generateMethodsForUnreachableProperties;
-        }
-
-    public INamedTypeSymbol TypeForBuilder { get; }
-    public string? BuildingMethodsPrefix { get; }
-    public bool? DefaultStaticCreator { get; }
-    public bool? ImplicitCast { get; }
-    public NullableStrategy? NullableStrategy { get; }
-    public bool? GenerateMethodsForUnreachableProperties { get; }
+    public INamedTypeSymbol TypeForBuilder { get; } = typeForBuilder;
+    public string? BuildingMethodsPrefix { get; } = buildingMethodsPrefix;
+    public bool? DefaultStaticCreator { get; } = staticCreator;
+    public bool? ImplicitCast { get; } = implicitCast;
+    public NullableStrategy? NullableStrategy { get; } = nullableStrategy;
+    public bool? GenerateMethodsForUnreachableProperties { get; } = generateMethodsForUnreachableProperties;
+    internal string? StaticFactoryMethodName { get; } = staticFactoryMethodName;
 }