+++ Tests/Buildenator.IntegrationTests.Source/Builders/EntityWithStaticFactoryMethodBuilder.cs
@@ -1,0 +1,10 @@
+﻿using Buildenator.Abstraction;
+using Buildenator.IntegrationTests.SharedEntities;
+
+namespace Buildenator.IntegrationTests.Source.Builders;
+
+[MakeBuilder(typeof(Entity), staticFactoryMethodName: nameof(SharedEntities.Entity.CreateEntity))]
+public partial class EntityWithStaticFactoryMethodBuilder
+{
+
+}